pkg: Move validations to operator package
