+++ pkg/prometheus/promcfg.go
@@ -30,7 +30,6 @@ import (
 
 	v1 "github.com/prometheus-operator/prometheus-operator/pkg/apis/monitoring/v1"
 	"github.com/prometheus-operator/prometheus-operator/pkg/assets"
-	"github.com/prometheus-operator/prometheus-operator/pkg/common"
 	"github.com/prometheus-operator/prometheus-operator/pkg/operator"
 )
 
@@ -219,56 +218,56 @@ func buildExternalLabels(p *v1.Prometheus) yaml.MapSlice {
 // validateConfigInputs runs extra validation on the Prometheus fields which can't be done at the CRD schema validation level.
 func validateConfigInputs(p *v1.Prometheus) error {
 	if p.Spec.EnforcedBodySizeLimit != "" {
-		if err := common.ValidateSizeField(p.Spec.EnforcedBodySizeLimit); err != nil {
+		if err := operator.ValidateSizeField(p.Spec.EnforcedBodySizeLimit); err != nil {
 			return errors.Wrap(err, "invalid enforcedBodySizeLimit value specified")
 		}
 	}
 
 	if p.Spec.RetentionSize != "" {
-		if err := common.ValidateSizeField(p.Spec.RetentionSize); err != nil {
+		if err := operator.ValidateSizeField(p.Spec.RetentionSize); err != nil {
 			return errors.Wrap(err, "invalid retentionSize value specified")
 		}
 	}
 
 	if p.Spec.Retention != "" {
-		if err := common.ValidateDurationField(p.Spec.Retention); err != nil {
+		if err := operator.ValidateDurationField(p.Spec.Retention); err != nil {
 			return errors.Wrap(err, "invalid retention value specified")
 		}
 	}
 
 	if p.Spec.ScrapeInterval != "" {
-		if err := common.ValidateDurationField(p.Spec.ScrapeInterval); err != nil {
+		if err := operator.ValidateDurationField(p.Spec.ScrapeInterval); err != nil {
 			return errors.Wrap(err, "invalid scrapeInterval value specified")
 		}
 	}
 
 	if p.Spec.ScrapeTimeout != "" {
-		if err := common.ValidateDurationField(p.Spec.ScrapeTimeout); err != nil {
+		if err := operator.ValidateDurationField(p.Spec.ScrapeTimeout); err != nil {
 			return errors.Wrap(err, "invalid scrapeTimeout value specified")
 		}
 	}
 
 	if p.Spec.EvaluationInterval != "" {
-		if err := common.ValidateDurationField(p.Spec.EvaluationInterval); err != nil {
+		if err := operator.ValidateDurationField(p.Spec.EvaluationInterval); err != nil {
 			return errors.Wrap(err, "invalid evaluationInterval value specified")
 		}
 	}
 
 	if p.Spec.Thanos != nil && p.Spec.Thanos.ReadyTimeout != "" {
-		if err := common.ValidateDurationField(p.Spec.Thanos.ReadyTimeout); err != nil {
+		if err := operator.ValidateDurationField(p.Spec.Thanos.ReadyTimeout); err != nil {
 			return errors.Wrap(err, "invalid thanos.readyTimeout value specified")
 		}
 	}
 
 	if p.Spec.Query != nil && p.Spec.Query.Timeout != nil && *p.Spec.Query.Timeout != "" {
-		if err := common.ValidateDurationField(*p.Spec.Query.Timeout); err != nil {
+		if err := operator.ValidateDurationField(*p.Spec.Query.Timeout); err != nil {
 			return errors.Wrap(err, "invalid query.timeout value specified")
 		}
 	}
 
 	for i, rr := range p.Spec.RemoteRead {
 		if rr.RemoteTimeout != "" {
-			if err := common.ValidateDurationField(rr.RemoteTimeout); err != nil {
+			if err := operator.ValidateDurationField(rr.RemoteTimeout); err != nil {
 				return errors.Wrapf(err, "invalid remoteRead[%d].remoteTimeout value specified", i)
 			}
 		}
@@ -276,13 +275,13 @@ func validateConfigInputs(p *v1.Prometheus) error {
 
 	for i, rw := range p.Spec.RemoteWrite {
 		if rw.RemoteTimeout != "" {
-			if err := common.ValidateDurationField(rw.RemoteTimeout); err != nil {
+			if err := operator.ValidateDurationField(rw.RemoteTimeout); err != nil {
 				return errors.Wrapf(err, "invalid remoteWrite[%d].remoteTimeout value specified", i)
 			}
 		}
 
 		if rw.MetadataConfig != nil && rw.MetadataConfig.SendInterval != "" {
-			if err := common.ValidateDurationField(rw.MetadataConfig.SendInterval); err != nil {
+			if err := operator.ValidateDurationField(rw.MetadataConfig.SendInterval); err != nil {
 				return errors.Wrapf(err, "invalid remoteWrite[%d].metadataConfig.sendInterval value specified", i)
 			}
 		}
@@ -291,7 +290,7 @@ func validateConfigInputs(p *v1.Prometheus) error {
 	if p.Spec.Alerting != nil {
 		for i, ap := range p.Spec.Alerting.Alertmanagers {
 			if ap.Timeout != nil && *ap.Timeout != "" {
-				if err := common.ValidateDurationField(*ap.Timeout); err != nil {
+				if err := operator.ValidateDurationField(*ap.Timeout); err != nil {
 					return errors.Wrapf(err, "invalid alertmanagers[%d].timeout value specified", i)
 				}
 			}
